Fix undefined case for serverProcess
+add comments

+++ app/main.js
@@ -7,13 +7,17 @@ const app = electron.app; // this is our app
 const BrowserWindow = electron.BrowserWindow; // This is a Module that creates windows
 
 let mainWindow; // saves a global reference to mainWindow so it doesn't get garbage collected
-let serverProcess;
+let serverProcess; // global reference to server child process for prod mode
 
 app.on('ready', async () => {
     // process is in dev mod or nah
     if (process.argv[2]) {
-        console.log('Running in dev Mode');
+        // dev mode runs a nodemon script to start the server
+        // see npm run watch
+        console.log('Running in dev mode');
     } else {
+        // prod/demo mode
+        // see npm run start
         console.log('Running in prod mode');
         serverProcess = fork('./src/server.js');
     }
@@ -50,7 +54,7 @@ function createWindow () {
 // when you close all the windows on a non-mac OS it quits the app
 app.on('window-all-closed', () => {
     // kill server child process
-    process.kill(serverProcess.pid);
+    if (serverProcess !== undefined) { process.kill(serverProcess.pid); }
     if (process.platform !== 'darwin') { app.quit(); }
 });
